Integrate next-international library and combine auth and internationalization middlewares
