--- app/(dashboard)/about/page.tsx
@@ -1,10 +1,0 @@
-import Title from "../../../components/UI/Title";
-import "../../../styles/About.css";
-
-export default function About() {
-  return (
-    <main className="about">
-      <Title titleName="Your Story Starts With Us." />
-    </main>
-  );
-}
