+++ app/[locale]/(dashboard)/layout.tsx
@@ -1,5 +1,5 @@
-import Header from "../../components/header/Header";
-import Footer from "../../components/footer/Footer";
+import Header from "../../../components/header/Header";
+import Footer from "../../../components/footer/Footer";
 
 export default function DashboardLayout({
   children,
