+++ app/[locale]/login/layout.tsx
@@ -1,4 +1,4 @@
-import ToggleThemeButton from "../../components/UI/ToggleTheme";
+import ToggleThemeButton from "../../../components/UI/ToggleTheme";
 
 export default function LoginLayout({
   children,
