--- app/login/page.tsx
@@ -1,11 +1,0 @@
-import LoginForm from "../../components/loginForm/LoginForm";
-import "../../styles/Login.css";
-
-export default async function Login() {
-  return (
-    <div className="login-container">
-      <h1>Login</h1>
-      <LoginForm />
-    </div>
-  );
-}
