+++ components/UI/Title.tsx
@@ -1,16 +1,12 @@
-"use client";
-
-import { useTranslation } from "react-i18next";
-
 interface TitleProps {
   titleName: string;
 }
 
 export default function Title({ titleName }: TitleProps) {
-  const { t } = useTranslation();
+  console.log(titleName);
   return (
     <div>
-      <h1>{t(`${titleName}`)}</h1>
+      <h1>{titleName}</h1>
     </div>
   );
 }
