+++ locales/client.ts
@@ -1,0 +1,14 @@
+"use client";
+
+import { createI18nClient } from "next-international/client";
+
+export const {
+  useI18n,
+  useScopedI18n,
+  I18nProviderClient,
+  useChangeLocale,
+  useCurrentLocale,
+} = createI18nClient({
+  en: () => import("./en"),
+  ka: () => import("./ka"),
+});
