+++ locales/en.ts
@@ -1,0 +1,63 @@
+export default {
+  // HEADER
+  home: "Home",
+  about: "About",
+  blog: "Blog",
+  contact: "Contact",
+  logOut: "Log Out",
+  theme: "Theme",
+  light: "Light",
+  dark: "Dark",
+  system: "System",
+
+  // FOOTER
+  quickLinks: "Quick Links",
+  information: "Information",
+  termsAndConditions: "Terms and Conditions",
+  privacyPolicy: "Privacy Policy",
+  newsletter: "Newsletter",
+  subscribeToOurNewsletter: "Subscribe to our newsletter:",
+  enterYourEmail: "Enter Your Email",
+  subscribe: "Subscribe",
+
+  // PRODUCTS PAGE
+  onlineStore: "Online Store",
+  searchProducts: "Search Products...",
+  search: "Search",
+  sortByPrice: "Sort by Price",
+  reset: "Reset",
+  addToCart: "Add to Cart",
+
+  //ABOUT PAGE
+  headline: "Your Story Starts With Us.",
+  learnMore: "Learn More",
+
+  // BLOG PAGE
+  readMore: "Read More",
+  archive: "Archive",
+
+  // CONTACT PAGE
+  contactUs: "Contact Us",
+  email: "Email",
+  phone: "Phone",
+  address: "Address",
+  name: "Name",
+  yourName: "Your name...",
+  yourEmail: "Your email...",
+  writeSomething: "Write something...",
+  subject: "Subject",
+  message: "Message",
+  send: "Send",
+
+  // PROFILE PAGE
+  userInformation: "User Information",
+  changePassword: "Change Password",
+  newPassword: "New Password",
+  confirmNewPassword: "Confirm New Password",
+  save: "Save",
+
+  // LOGIN PAGE
+  username: "Username",
+  password: "Password",
+  login: "Login",
+} as const;
