+++ locales/ka.ts
@@ -1,0 +1,63 @@
+export default {
+  // HEADER
+  home: "მთავარი",
+  about: "ჩვენ შესახებ",
+  blog: "ბლოგი",
+  contact: "კონტაქტი",
+  logOut: "გასვლა",
+  theme: "თემა",
+  light: "ღია",
+  dark: "მუქი",
+  system: "სისტემა",
+
+  // FOOTER
+  quickLinks: "სწრაფი ლინკები",
+  information: "ინფორმაცია",
+  termsAndConditions: "წესები და პირობები",
+  privacyPolicy: "კონფიდენციალურობის პოლიტიკა",
+  newsletter: "სიახლეები",
+  subscribeToOurNewsletter: "გამოიწერეთ სიახლეები:",
+  enterYourEmail: "შეიყვანეთ თქვენი მეილი",
+  subscribe: "გამოწერა",
+
+  // PRODUCTS PAGE
+  onlineStore: "ონლაინ მაღაზია",
+  searchProducts: "მოძებნე პროდუქტები...",
+  search: "ძიება",
+  sortByPrice: "ფასით დალაგება",
+  reset: "დალაგების გაუქმება",
+  addToCart: "კალათაში დამატება",
+
+  //ABOUT PAGE
+  headline: "შენი მოგზაურობა იწყება აქ.",
+  learnMore: "გაიგე მეტი",
+
+  // BLOG PAGE
+  readMore: "წაიკითხე მეტი",
+  archive: "არქივი",
+
+  // CONTACT PAGE
+  contactUs: "დაგვიკავშირდი",
+  email: "მეილი",
+  phone: "ტელეფონი",
+  address: "მისამართი",
+  name: "სახელი",
+  yourName: "შენი სახელი...",
+  yourEmail: "შენი მეილი...",
+  writeSomething: "დაწერე ტექსტი...",
+  subject: "თემა",
+  message: "წერილი",
+  send: "გაგზავნა",
+
+  // PROFILE PAGE
+  userInformation: "მომხმარებლის ინფორმაცია",
+  changePassword: "შეცვალე პაროლი",
+  newPassword: "ახალი პაროლი",
+  confirmNewPassword: "გაიმეორეთ ახალი პაროლი",
+  save: "შენახვა",
+
+  // LOGIN PAGE
+  username: "მომხმარებელი",
+  password: "პაროლი",
+  login: "შესვლა",
+} as const;
