+++ locales/server.ts
@@ -1,0 +1,6 @@
+import { createI18nServer } from "next-international/server";
+
+export const { getI18n, getScopedI18n, getStaticParams } = createI18nServer({
+  en: () => import("./en"),
+  ka: () => import("./ka"),
+});
