+++ next.config.mjs
@@ -3,11 +3,6 @@ const nextConfig = {
   images: {
     domains: ["cdn.dummyjson.com"],
   },
-  // i18n: {
-  //   locales: ["en", "ge"],
-  //   defaultLocale: "en",
-  //   localeDetection: false,
-  // },
 };
 
 export default nextConfig;
