+++ package.json
@@ -9,11 +9,10 @@
     "lint": "next lint"
   },
   "dependencies": {
-    "i18next": "^23.11.2",
     "next": "14.1.4",
+    "next-international": "^1.2.4",
     "react": "^18",
-    "react-dom": "^18",
-    "react-i18next": "^14.1.1"
+    "react-dom": "^18"
   },
   "devDependencies": {
     "@types/node": "20.12.7",
